requirements.txt
Django==1.11.7
django-ckeditor==5.3.1
django-js-asset==0.1.1
olefile==0.44
Pillow==4.3.0
pytz==2017.3
